Created Primesuptocsv
This is Matlab code. It calculates all of the prime numbers up to the limit 'Lim' and writes them to a comer separated value file, 'prime.dat'. Problems: Numbers ~ 10^5 are written in exponential form in the output file. This is obviously of no use as the precision is needed down to the last digit.